updateDelete

+++ deleteCharacter.js
@@ -1,32 +1,14 @@
-function myConfirm() {
-  let result = confirm("Want to delete?");
-  if (result==true) {
-   return true;
-  } else {
-   return false;
-  }
-}
-
-(() => {
-    const input = document.querySelector("#tpl-card")
-
-    document.querySelector('.btn-danger').addEventListener("click", async () => {
-        console.log("hello");
-        /*
-        try {
-            const response = await fetch("http://character-database.becode.xyz/characters"+input.value, {
-                method: 'DELETE'
-                
-            });
-           
-            if(!response.ok) throw new Error(response.statusText)
-            const getAll = await fetch('http://character-database.becode.xyz/characters')
-            const data = await getAll.json()
-            console.log(data)
-        } catch (err) {
-            alert(err)
-        } */
-    });
-})();
-
-
+document.getElementById('delete').addEventListener("click", async () => {
+    let action = confirm('Are you sure you want to delete this character');
+  
+    if (action != true) {
+      console.log("action aborted nothing deleted");
+    } else {
+      const result = await fetch('https://character-database.becode.xyz/characters/${characterId}', {
+              method: 'DELETE',
+          });
+          window.location.href = "index.html"
+          console.log("Deleted")
+    }
+  }); 
+
